Unsubbed from TickTimerService and added Hide Time
I unsubscribed from the TickTimerService since I don't need to do anything at any temporal interval. I only want the time when I shake the watch. So the accel subscription is registered on init() and when an event is detected, it will update the time and show the time layer. 5 seconds later, a timer will run a callback to make the time invisible again (as the time is not being updated regularly).